scripts/requirements.txt
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-auth==2.6.0
google-cloud-core==2.2.3
google-cloud-storage==2.2.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
idna==3.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
requests==2.27.1
rsa==4.8
six==1.16.0
toml==0.10.2
urllib3==1.26.8
